Add support for specifying a login name
The remote account used for configuration may not have the same name as
the current local user and probably _should_ not anyway. Look for a
username in the host argument, -u option, ACT_SSH_USER environment
variable, in that order, or fall back to the local user name obtained by
id(1).

Signed-off-by: Alex Minghella <a@minghella.net>