Add log socket handling functions
NOTE: This is a work in progress.

Currently, there's no way to write to the log file from modules running
on the remote, which means that we have a log file containing only what
we did on the local side of the connection. It would be good to receive
log information from modules.

It would be nice if we could open a file descriptor that ssh(1) would
pass to the remote command, but it only handles stdin(3), stdout(3), and
stderr(3) like that.

There are a few solutions to this:

1. Write logs to stderr(3) and redirect to the log file.
2. Write logs to a file on the remote and copy its content back over and
   append it to the log after we're done.
3. Forward a port using the "-R" option of ssh(1) and have modules write
   logs to that.

Option 1 is out because using stderr(3) specifically for log output
means restricting modules to a single output file descriptor, stdout(3).
Redirecting stderr(3) from ssh(1) in general is probably a bad idea to
begin with, since it uses it for various kinds of diagnostic output.

Option 2 causes us to write a log file out of order. How badly out of
order depends on when we append the remote log to the local log.

This commit introduces functions to facilitate option 3; we forward a
Unix socket so that we can receive log information from modules running
on the remote. After we forward the socket, we use nc(1) to listen on it
and write everything it receives directly to the log file.

Modules on the remote can bind to this socket to write log information;
how they do so is up to them.

Signed-off-by: Alex Minghella <a@minghella.net>